correct mistakes in Rate component
