+++ src/pages/comparison/components/profile.component.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import '../comparison.scss';
 
-import { ProfileInterface } from '../interfaces/pet.interface';
+import { ProfileProps } from '../interfaces/pet.interface';
 
-export const Profile: React.FC<ProfileInterface> = props => {
+export const Profile: React.FC<ProfileProps> = props => {
   return (
     <div className="pet-block">
       <img className="pet-image" src={props.imgUrl} alt={props.breed} />
