+++ src/pages/comparison/components/rate.component.tsx
@@ -1,28 +1,13 @@
 import React from 'react';
-import './rate.css';
+import './rate.scss';
+import clases from './rate-logic.service';
 
-import { RateInterface } from '../interfaces/pet.interface';
+import { RateProps } from '../interfaces/pet.interface';
 
-export const Rate: React.FC<RateInterface> = props => {
-  const { rate } = props;
+export const Rate: React.FC<RateProps> = props => {
 
-  const clases: string[] = ['gy', 'gy', 'gy', 'gy', 'gy']; // gy - grey
-
-  if (rate === 1) {
-    clases[0] = 'r'; // r - red
-  } else if (rate < 4) {
-    for (let i = 0; i < rate; i++) {
-      clases[i] = 'y'; // y - yellow
-    }
-  } else {
-    for (let i = 0; i < rate; i++) {
-      clases[i] = 'gr'; // gr - green
-    }
-  }
-
-  let key = 0;
-  const items = clases.map(el => {
-    return <div key={key++} className={el}></div>;
+  const items = clases(props.rate).map((el, index) => {
+    return <div key={index} className={el}></div>;
   });
 
   return <div className="rate-bar">{items}</div>;
