--- src/pages/comparison/components/rate.css
@@ -1,25 +1,0 @@
-.rate-bar {
-    display: flex;
-}
-.rate-bar div {
-    margin: 5px;
-    height: 15px;
-    width: 15px;
-    border-radius: 100%;
-    background-color: #E4E4E4;
-}
-.rate-bar .r {
-    background-color: #FF453A;
-}
-
-.rate-bar .y {
-    background-color: #FF9F0A;
-}
-
-.rate-bar .gr {
-    background-color: #30D158;
-}
-
-.rate-bar .gy {
-    background-color: #E4E4E4;
-}
