+++ src/pages/comparison/components/rate.scss
@@ -1,0 +1,27 @@
+.rate-bar {
+  display: flex;
+
+  div {
+    margin: 5px;
+    height: 15px;
+    width: 15px;
+    border-radius: 100%;
+    background-color: #e4e4e4;
+  }
+
+  .r {
+    background-color: #ff453a;
+  }
+
+  .y {
+    background-color: #ff9f0a;
+  }
+
+  .gr {
+    background-color: #30d158;
+  }
+
+  .gy {
+    background-color: #e4e4e4;
+  }
+}
