+++ src/store/reducers/root.reducer.ts
@@ -1,4 +1,5 @@
 import { combineReducers } from 'redux';
+
 import comparisonReducer from './comparison.reducer';
 
 export default combineReducers({
